Change percent field to double

+++ MediportaTest/Models/Tag.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel.DataAnnotations.Schema;
+
 namespace MediportaTest.Models;
 
 public class Tag
@@ -5,5 +7,5 @@ public class Tag
     public long Id { get; set; }
     public string Name { get; set;}
     public long Count { get; set;}
-    public decimal Percent { get; set; }
+    public double Percent { get; set; }
 }
